[pc ui] Adds support to use the pc filter in saved views
In order for this to work, the tree should be able to be initialized (its selected nodes checked) based on the filter criteria, and the product characteristic
s summary should be able to be obtained using the criteria too. The criteria itself did not have enough information to tbuild the product characteristics summ
ary, so the filter internal value is being stored along with the criteria.

The tree initial selected nodes are checked based on this internal value. This is done in the dataArrived function, before the data has arrived it is not poss
ible to do it, because the nodes are not added to the tree until then.