Fixes issue 23870: Filters work properly with property columns
In the buildFieldClause method of the AdvancedQueryBuilder class there was a bug that prevented filtering from working in the case where the size of the properties array was longer than 2. This happened because at one point the useFieldName string was being truncated from the first occurrence of the point, when it should have been truncated from the last.

--HG--
extra : rebase_source : 52322eb7102953c0c88b09cd47e0192ec21dab06