Fixes issue 23889: No extra datasource calls are done by selectors
Pick lists and grids use a property called drawAllMaxCells, that is used to determine how many cells should be fetched from the datasource, its default value is 250. If after making a datasource call smartclient finds out that the number of retrieved cells is lower than drawAllMaxCells, it will do another datasource call, and repeat this process until the number of cells fetched are higher than drawAllMaxCells or until all cells have been fetched.

This is not the intented behaviour, we want the rows to be fetched when the user scrolls down and is out of cached rows. That is why in the definition of OBViewGrid the drawAllMaxCells is set to 0. To fix this issue this property is also set to 0 in the selector pick list and grid.

--HG--
extra : rebase_source : 024d491f6756362088a7e58a6959cd6f35656ac6