Fixes issue 23890: The same hql is generated in the grid and in form view
The problem was that for that particular selector (Product selector), the generated displayfield string was 'product$product$name'. This is not correct, because the Product entity does not have a product property, the correct displayfield string is product$name. This has been fixed in the getDisplayFielName method of the FKSelectorUIDefinition class. A check has been added to make sure that the property used in the display field is part of the referenced entity. If it is not, then the displayFieldName is returned directly, instead of trying to append it to the referenced property.

--HG--
extra : rebase_source : 1b2279710b1a9ca95c49b593451c5afa55bd4d45