Fixes issue 23078: Child tabs use parent ad_module_id as default
The module column of child tabs should not be defaulted to the module flagged as default, but to the module of the record parent. This has been fixed for the module columns for the Menu - Menu Parameters and the Reference - User Interface Definition tabs.

--HG--
extra : rebase_source : 9e8e244d6d5b1c180017ff0b63c9153bee53e8b0