Fixes issue 23975: Number handling works using ',' as decimal separator
In the fix of the issue 23758 [1] it was wrongly assumed that the function getValue() of a OBNumberItem was going to return a javascript number. When the an existing record has just been opened getValue() will return a number, but when the user edits it, it will return the string representation of the number using the format specified by the user.

Now this has been taken into account, so the value is treated differently if getValue returns a string or a javascript number.

[1] https://code.openbravo.com/erp/devel/pi/rev/2e4eccaa3cc8fc8ac434cb0d386166558d1ee686

--HG--
extra : rebase_source : 88573fd7e6012f87e6f979b75041ad4dc254613a