Fixes issue 23991: table_name column of ad_table table is mandatory again
At one point durint the development of the Datasource Based Tables project the table_name column of the ad_table table was set not mandatory, because this field was meant to be left empty for tables based on datasources. Later in the project it was decided to use the column 'name' (the name column of the Table table) to set the 'table_name' column for datasource based tables. This was we avoided having to change the table identifier (table_name).

After this change the table_name column must be mandatory in the database.