Fixes issue 23996: An event handler has been replaced with database constraints
The event handler TableDataOriginEventHandler has been replaced with database constraints. This is a better approach because it is simpler and easier to maint
ain.

The name of two of the messages has been updated so that it is the name as the constraint that throw them. The message used to warn the user that the table_name must be set if the data origin is 'Table' has been removed because now the table_name is mandatory again in the database.

--HG--
extra : transplant_source : t%28%95%BC%82%E3%9B%C8%13C%88%9A%91O%A7W%BCe%3D%7E