Fixes Issue 24021, Fixes Issue 23756 : Inhomogeneous behavior of some filters
Some filters after refresh or process completion did not retain the filter properly and other records were shown.
In refreshGrid, the newRecordsToBeIncluded array contains rows that do not adhere to the current filter.
The array is now compared with the data in the grid and the records are added only if its present in the current grid filter.
This change will not affect records opened using openDirectTab. Those records will be displayed even if not part of current filter.