Fixes issue 23990: XML is imported properly using Initial Org Setup
This is what was happening when an object with id #id was being imported from a dataset to a organization. #id is the id of the object in the dataset
- The first time the object is imported to a organization (org1), the original #id is used
- When the same record was being imported to a sister organization (org2), a query for that object with the id #id was done. The organization filter was not being applied, so the object that was imported in the previous step was found. This object was related with other objects that belonged to org1, but was being referenced too from other objects that were belonged to org2. org1 is not part of the natural tree of org2, so an exception was thrown.

This has been fixed by adding a parameter to the resolve method that will force to apply the organization filter. The default value of this parameter is false, and it will only be true then is called from the InitialSetupUtility class.

--HG--
extra : rebase_source : 12df422e3a2d966a91b236b7ce40a47846936af4