Fixes issue 23856: Records are not duplicated when saving repeatedly
Now the standard view datasource checks if there is an ongoing save transaction before saving a record. If the is, it does not calls the datasource again, because that would result in the record being duplicated in the database.

--HG--
extra : rebase_source : afb54a45cdee786502f544fac1ef153d85a59231