Fixes issue 24120: Widgets embedded in tabs are supported again
In this changeset [1] some logic was added to restrict the access to the widgets. One of the checks consisted in throwing an exception if the widget class did not have any widget instance class associated. This makes sense for workspace widgets, but widgets embedded in tabs do not have widget instance classes.

Now, this check is only done if the widget is not embedded in a window accessible by the user.


[1] https://code.openbravo.com/erp/devel/pi/rev/08d5a13b722bbc496c640a976d5bb319a1db44bc

--HG--
extra : rebase_source : 2a600f9f17f1461c52dded93ca52b0cb286d1388