Fixes issue 24152: P&E grid rows are sent to the Java Process
In a P&E grid, if all the grid records are cached they will be sent to the server (there are some java processes that need to access all the records). To check if all the records were cached we used to check grid.data.cachedRow, but that does not work always. Now to check this we compare the number of cached rows with the resultSize property of the data. If the number of cached rows is lower than the resultSize property (maximum number of rows that will be fetched in each datasource call), that means that there are no more records pending to be cached.

--HG--
extra : rebase_source : d0e3dab5c73b8ae5dde9758533f3f8b78d75c35e