Fixes Issue 0024092: Newly initiated client does not appear in profile list
When a new client is created, an information is shown which provides a link to reload the context so that the newly created client will appear in the user profile widget.
When a client is deleted from delete client and accessed immediately from user profile widget an error is thrown instead of null pointer exception.