Fixes issue 24181: Displayed Value flag in Table Reference works
The Displayed Value flag was not being taken into account in the DataToJsonConverter class, so it was not being shown in the grid. To fix this, the displayedV
alue property has been added to the RefTable base class. This property is checked in the isDisplayedValue method of the Property class which is used in the Da
taToJsonConverter to check if the value (searchKey property) of the entity should be appended to its identifier.