Fixes issue 24231: It is possible to apply a dataset in two organizations
In the fix of the issue 23990 [1] a new parameter was added to the resolve method of the EntityResolver class, that had to be set to true when it was called when datasets are being applied. This code was included in the process called when an organization is created with datasets, but not in the Enterprise Module Management process.

[1] https://code.openbravo.com/erp/devel/pi/rev/83447d2c67e21ac06ca92c1d2ec432d5e9e2656a