[rest webservices with selected properties] First commit
This changeset implements the functionality that enables to specify the list of properties that will be exported in a XML Rest Webservice call (this functionality was already implemented for Json Rest Webservices).

Changes done:
- DalWebService.java: If the _selectedProperties is included in the request parameters the string is parsed to obtain two maps (propertiesToBeFeched and childPropertiesToBeFeched) that will contain the list of child and non child property for each entity being exported (i.e. when an entity is beign exported its child entities can also be exported in the process). Then these two maps are passed to the EntityXMLConverter.
- EntityXMLConverter.java: Filters out the properties not included in the propertiesToBeFetched or childPropertiesToBeFetched maps, but only if they are not empty. If they are empty (because the _selectedProperties parameter was not included in the URL) then all properties are returned (the child properties will only be returned if the includeChildren parameter is not set to false).

Example:
Let's suppose that someone wants to retrieve the list of all windows, including its tabs and its fields. Only these properties are required:
- ADWindow: id, name
- ADTab: id, name, tabLevel
- ADField: id, name, description

The request to the xml webservice should be like this:

http://localhost:8080/openbravo/ws/dal/ADWindow/100?_selectedProperties=id,name,aDTabList,aDTabList.id,aDTabList.name,aDTabList.tabLevel,aDTabList.aDFieldList,aDTabList.aDFieldList.id,aDTabList.aDFieldList.name,aDTabList.aDFieldList.description

'id', 'name' and 'aDTabList' are properties of the ADWindow entity. The aDTabList property is a child entity of ADWindow. If no more properties were specified, then only the id and the identifier of each tab would be returned.
'aDTabList.id','aDTabList.name','aDTabList.tabLevel' and 'aDTabList.aDFieldList' are properties of the adTabList entity.
'aDTabList.aDFieldList.id','aDTabList.aDFieldList.name' and 'aDTabList.aDFieldList.description'

When the _selectedProperties parameter is parsed, these two maps will be generated:
- properties: a map with the list of not child properties to be exported for each entity:
      {_top=[id, name],aDTabList=[id, name, tabLevel], aDTabList.aDFieldList=[id, name, description]}
- childProperties: a map with the list of child properties to be exported for each entity:
      {_top=[id, aDTabList],aDTabList=[id, aDFieldList]}