Fixes issue 24390: Only System Admin can define the module of preferences
Only the role System Admin should be able to define the ad_module_id column of the preferences (only the preferences with ad_module_id defined can be exported). This has been achieved by making that field visible only for that user (by default that field is always empty).

A server side check is not going to be done, because it is risky and it could result in proper preferences not being exported.

To fix the preference data of existing faulty preferences (preferences defined by a role other than system admin with invalid visibleat_role_id, visibleat_client_id or visibleat_org_id values) this query could be executed:

update ad_preference
set ad_module_id = null
where (visibleat_role_id is not null and visibleat_role_id != '0')
or (visibleat_client_id is not null and visibleat_client_id != '0')
or (visibleat_org_id is not null and visibleat_org_id != '0')