Fixes issue 24228, issue 24229: Avoids not needed datasource
Some not needed datasource calls were being made when a record was created in grid or form view. It made no sense to ask the datasource for the data of the child tabs of the new record, we can be sure there are none. To fix this, instead of calling the datasource the data of the child grid is emptied manually, and then the emptyMessage of the grid is updated.

--HG--
extra : rebase_source : 317e783ec4558dfb166ddd4affcd0764f1449b6e