Fixes issue 24461: Date fields can be modified from callouts more than once
Due to a change done in this changeset [1], modifying a date field from a callout only worked the first time, from then on the value was not updated. This is caused because when a date time is modified from a callout its _textChanged property is set to true, but never set back to false (this property should only be true the time while this field is being updated). Setting it back to false after modifying its value from a callout fixes the problem.

[1] https://code.openbravo.com/erp/devel/pi/rev/4060f2b5e6bd2429d831c5e30cc4d45892efcd96

--HG--
extra : transplant_source : %AD%00%C4%DD%B9%A1%CFW%D6%FB%E1e%F3%8E%3Cq%CCS%C2%B0