Fixes issue 24491: DateTime fields can be modified from callouts
The problem was the following:
- The current value of the datetime field is, i.e., 05-06-2013 10:00:00
- The callout tries to change it to another datetime, but not modifying the date part, i.e. 05-06-2013 11:45:00
- In order to check if the value of the field needed to be changed, the function isc.Date.compareLogicalDates(newValue, oldValue) was used. This function only
checks the date part of the value. In this case the date part is the same, so the change was not applied.

To fix this, datetime fields use isc.Date.compareDates (which compares both the date and the time parts) instead of isc.Date.compareLogicalDates.

--HG--
extra : rebase_source : eb21c6ac57c4be2cac7dafa7a8f5a06adf3fe6cf