Fixes Issue 0024574: Filter is not working fine in some special cases
When list items are filtered by selecting the option from dropdown, the filter applied is filtering using contains. Hence if the filter is 'test A', 'test B A' will also be shown.
When selecting values, == is appended to each selected item to filter the exact value.
When showing picklist, == is trimmed from value so that the values are selected properly in the dropdown.