Fixes issue 24566: The proper summary function description text is shown
The problem was that the description text was being set in the field of the OBViewGrid, but then read from the field of OBViewGrid_summaryRow. To fix this, the description text is going to be read from the same field that was used to set it.

--HG--
extra : rebase_source : 617352f3302f42a89795f2eb372b37c94e61afe3