Fixes Issue 0024637: Display Logic does not work in tabs located in windows that have several tabs referencing the same entity
When we have more than one entity referenced in the same window, the header reference is overwritten in the hash-map with the tab referencing the same entity.
Overriding is prevented and only if entity is not found in the hash map, it is added.

--HG--
extra : rebase_source : 35a93f61972f663b0cc8def35abaf5adc57b98d2