Fixes Issue 0024698: When opening a window two request to save the view on recent view section are send
When launching a menu through quick launch, two calls to save the Recent view is made.
One is happening in the quick launch and other in the default openView method.
Removed the code to add to recent view in quick launch.
The same code is accessed when creating a new record through create new record so the issue is fixed there too.