Adaptive Filtering Project
This changeset enables the use of adaptive filtering in Openbravo OBViewGrids. The adaptive filtering allows to do client side filtering and sorting of the grid records if the dataset is small enough. Not only the filtering and sorting is almost immediate, but the server load is reduced severily as the server is no longer required to do those tasks.