Fixes Issue 0024726: Problems with Selectors if you migrate from Openbravo 2.22
In the fix for the issue 20352 [1], the condition to replace the numbers only if its a 32 characters UUID was introduced.
But there are ID's in the system that is upto 10 characters, but all numbers. So it does not replace the numbers for these cases,
so instead of the text, the ID's were shown. Now added condition to check the following:
1. uuid's length should be less than 10 or equal to 32.
2. If uuid's length is less than 10, it should contain only numeric values

[1] https://code.openbravo.com/erp/devel/pi/rev/0a1670b979e2c0d42e97e8c8b045b4a760df9a28

--HG--
extra : rebase_source : 7e376bbe7e932cbfde0ee861d649a98e62dbc4e8