Fixes issue 24741: Row saved by pressing a tab for picking a selector value
The problem was that the request to the datasource to save the new record was being done before the response of the FIC (it was called because picking a business partner triggers a callout to select its location) was received and processed. This has been fixed by delaying the execution of the cellEditEnd function until the FIC response has been processed.

--HG--
extra : transplant_source : %918%CD%FB%CB%9B%AD%FC%EA%F1%A0%EBT%B4%BE%87%FE%C6%EB5