Fixes Issue 0024692: Double request to the datasource is done when a numeric field is in the first position of the grid
In numeric fields, then in the blur event, the filters are applied again if its not applied through performGridAction.
But this was resulting in double datasource calls, if the numeric field is the first field of the grid.
Added validation to performGridAction only if the numeric field is not the first field of the grid.