Fixes Issue 0024692: Double request to the datasource is done when a numeric field is in the first position of the grid
The filter in blur event needs to be applied only if the value is not undefined.
In case there is value in the filter and then the value is cleared,
still the filter will be applied because the value returned will be null and not undefined.
Also the additional datasource call will be avoided as in that case, the value is undefined.