Fixes issue 24924: Changes are applied to the grid with adaptive filtering
When a record was updated in form view with the record filtered using adaptive filtering, the changes were not being applied to the grid. This happened because the changes were being applied to the allRows property of the ResultSet, but not to the localData property, which is what contains the data of the filtered records.

--HG--
extra : transplant_source : %2B%5C%09%BE%D0%87%93%CA%A2%A4%D6Gx%DE%24%E5%DA%ACe%8C