Fixes bug 24954: Inserting a new recort not working a previous insert error
The problem was that the view._savingNewRecord was not being reset after the a new record request to the datasource coming from the grid resulted in an error. Now this case in handled in the editFailed function of ob-view-grid.

--HG--
extra : rebase_source : 47ea2b8287d6e60504a2128009be8b4688f6c951