Fixes bug 24954: Inserting a new recort not working a previous insert error
The problem was that the view._savingNewRecord was not being reset after the a new record request to the datasource coming from the grid resulted in an error. Now this case in handled in the editFailed function of ob-view-grid.

--HG--
extra : transplant_source : %DD%9E%AB%1A%DE%A0n%28Mjw%E9%EB%AC%F3%F4Z-m%92