Fixes issue 24985: Audit data values are not shown in the form view
One of the changes introduced by the grid datasource project [1] was that the grid does not load all the properties for each row, only the ones that are visible in the grid (plus a few other required ones). When a record is opened in form view there is call to the FIC in EDIT mode to retrieve all the record values. The problem is that the audit values are not being returned in that FIC call, because there are no fields for those values.

To fix this, the audit values have been added to the list of mandatory properties, so they will be loaded to the grid even if they are not being shown.

[1] https://code.openbravo.com/erp/devel/pi/rev/94d0fb151919

--HG--
extra : rebase_source : ff1886b52f0755f8e5eeb6cd23544650278a9577