Fixes issue 24951: Grid always fetches properties used in subtabs display logic
In order for the display logics of a subtab to be evaluated properly, the properties that belong to an ancestor tab must be loaded in that tab.

A new utility method called getTabSubtabs has been added to the KernelUtils class.

--HG--
extra : rebase_source : 135f9a5af80ab17c0695441118a061446a605a58