Fixes bug 24995: Moving to the next record in the grid with 'Tab' does not work
In this changeset [1] it was wrongly assumed that the newRow property represented whether the record being edit was new, but it represents if the grid is about to move to a new line. To check if the row is being created, this.getEditForm().isNew should be used.

[1] https://code.openbravo.com/erp/devel/pi/rev/9138cdfbcb9badfceaf1a0eb54b4be87fec6eb35

--HG--
extra : transplant_source : _%9A%1B%F7%00%5C%EB%E1%08%1D%87%F3%C2%BA%21%BF%CA%FB%11%A3