Fixes issue 25215: All buttons are included in the grid datasource response
Initially only the buttons with label values were considered required for the grid. This was not enough, because there are buttons with docAction that need to have their value present when the button is pressed in order to compute its combos properly.

--HG--
extra : rebase_source : 56a956d39a40250d011915b636270a684179c9cd