Fixes issue 25307: Tab display logic works with preferences
For computing the display logic of a tab, the context info of its parent tab is used. But if a preference is used, it is stored as a session attribute in the subtab, so it was not being included in the context.

--HG--
extra : transplant_source : %9A%DA%C4%AFC%05%23%CF%87%EC%1D%1C%D5%83%25%0DT%D7%A2%05