Related to issue 24615: implemented double cash criteria
For purchase flows the cash vat can be defined by the business partner (vendor) OR the organization.
If both the Cash VAT and the Double Cash Criteria flags are set to Yes, the Cash VAT regime will be automatically applied for the purchase flow too.

The sales flows only depend on the Organization's Cash VAT flag.