Fixes issue 25481: Time fields are properly parsed
When a row was being opened in view form, an error in the log was shown if a time field that was not displayed in the grid is present in the view form. This happened because in the moment when its value was being parsed it also contained the date.

This has been fixed by making sure that when the time is parsed only the time part is used.