Fixes issue 25484: It is possible to enter numbers in the product selector
When a number was entered in the product selector (Purchase Invoice window, Lines tab), it was being automatically deleted. This happened because it was being
detected as an UUID, and we try to avoid showing IDs whenever is possible.

To avoid this, the original value is restored in the redraw() function of ob-form-view in this case:
- The focused form item is a foreign key
- It has lost its displayed value
- The original displayed value was a number

This is not going to intefere with the rest of the cases where the UUID should actually be not shown. Those cases are handled in functions other than redraw()

--HG--
extra : transplant_source : %92%9C%A6%80%10%12%99%DA%2A%01%87%7C%19%3E%99%0D%FCM%FC%84