Fixes issue 24580: Casing was being wrongfully ignored in some filters
When filtering a column that pointed to a table with a multiple identifier, sometimes the casing was being ignored when it shouldn't. All the properties that conform the identifier should be taken into account, not just the first one.

--HG--
extra : rebase_source : 08f2fe5f6a8a5306f2abc22378ba2936107423a4