Fixes issue 25633: Symbol of default operator is not shown
The problem was that even though the operator of the column was iStartsWith this.getOperator() returned iContains. That meant the iContains was considered the
default operator. The default operator symbol is not shown in the filter, that is why the '~' (iContains symbol) was removed after filtering. Then, if the grid
 was filtered again there was no explicit operator, so the operator defined in the grid configuration was used.