Fixes bug 25666: Columns are not wrongly updated when chaging rows in the grid
This issue happened when the user clicked on a record (this triggers a FIC call) and then clicked on another record before the FIC response was received. What
was happening is that the FIC response was applied on the currently selected row, not on the row that trigger the FIC call. This has been fixed by checking bef
ore the FIC response is processed if the selected row has changed.