Query Timeout Feature
Allows to define a query timeout for hibernate queries, criterias (OBCriteria) and for sql statements. New methods are added to the classes generated based on xsql files that enable defining a query timeout. The user will have to define in a properties file the query timeout for each different query time. The name of the property must be bbdd.queryTimeout.<queryType>

For instance, to add a timeout to a hibernate query this command can be used:

QueryTimeOutUtil.getInstance().setQueryTimeOut(qry, "myQueryType");

Then in the Openbravo.properties file, the following line should be added.

bbdd.queryTimeout.myQueryType=2

In that case, if the query takes longer than 2 seconds to execute it will be canceled and an exception will be thrown.