Fixes issue 25581:  Display logic is not working fine using auxiliary input
The problem was that the display logic was using an auxiliary input defined for the parent tab, but its value was being wrongly overwritten by the subtab (the tab being hidden/shown). I
t has been fixed by adding the subtab session properties only if they are not an auxiliary input in the parent tab.