Fixes issue 25710: Big numbers are not properly managed in amount fields
The problem was that the OB.Utilities.Number.OBMaskedToOBPlain function was removing the '0's after the decimal dot even if the number was being expressed in decimal notation. This way, '1.12E10' was being converted to '1.12E1'.

This has been fixed by not removing the lasting '0's if the number is expressed in decmial notation.

--HG--
extra : rebase_source : 91ff80eccefda5db3ba5f5c0d1bea41b50a3ed97