[Connection Pool Extension] Changes how xsql classes are generated
In the first draft of the Query Timeout project, a new method declaration was added for each method defined in the xsql file, to allow calling that method passing the query type as an argument.

Most of the times those methods are going to be used without passing the query type as an argument, so a lot of overhead was being created. This has been fixed by adding the extra method declaration only if the useQueryProfile parameter is set in the declaration of the method in the xsql file.

--HG--
extra : rebase_source : 8d4441921b54f584a120cd20fa9dce39ceb1efbe