[Tree Extension] Reparentings nodes in the Menu ADTree is different
When a node is moved in an ordered ADTree, the following happens:
- The moved node gets a new sequence number
- Its new siblings with a higher sequence number have their sequence number increased by 10.

When a node is moved in the Menu ADTree, the behaviour is a little different. Not all the new siblings will be updated. The siblings are updated until one of them belongs to a module not in development, then no more nodes are updated.