fixed bug 25765: duplicated records can appear in the grid when it is refreshed
When the grid is refreshed, viewGrid.targetRecordId is set which is placed in the context.params._targetRecordId. This parameter should not be sent when requesting next page. The code in charge of removing it, is in viewGrid.grid.getFetchRequestParams which is called from viewGrid.fetchRemoteData, the problem is the result is not correctly set back in context.params.