[Tree Extension] Adds IsMainTree column to AD_TABLE_TREE
A table can define several trees, but only one of them can be the main tree. The main tree is used when a node is added to the tree or removed from it. The add or remove methods of the datasource of the main tree will be executed when the node is created or deleted.

If one of the trees of a table uses the ADTree tree structure, that must be the main tree.