[Tree Extension] Defines ROOT_NODE_DB and ROOT_NODE_CLIENT
The ROOT_NODE constant, defined in TreeDatasourceService, was used to set the parent_id of the root nodes (being root nodes, they actually don't have a parent). It was not possible to use '0' as ROOT_NODE in the client, because the AD_ORG table has a root node (* organization) with nodeId = '0'. The parent_id of this node would have to be '0', which results in a loop.

Therefore, '-1' is going to be used to represent the parent of root nodes in the client (ROOT_NODE_CLIENT), while '0' is still going to be used to represent the parent_id of root nodes in the database (ROOT_NODE_DB).