[Tree Extension] Takes into account ROOT_NODE_DB in the moveNode method
Now that the ROOT_NODE_CLIENT property is equals to -1, if a node is moved to the root, its parentId will be -1 in the moveNode method. Depending of whether the tree is the org tree or not, this parentId will have to be converted to ROOT_NODE_DB or to null.