[Tree Extension] Adds UpdateADTrees modulescript
This modulescript fills in the ad_table_id field of the ad_trees that do not have it defined. There is a 1:1 relation between the treetype column of ad_tree and the treetype column of ad_table, so this relation is used to set the ad_table_id column of the ad_tree table.