[Tree Extension] Changes in InitialClientSetup
One step of the InitialClientSetup process creates the ADTrees for the new client. The list of ADTrees to be created used to be obtained using the ADTreeType Type list. This list is no longer going to be maintained. Instead, now the list of ADTrees to be created is obtained using the list of tables that have a tree defined with the ADTree tree structure.

Some more changes have been done to take into account that now ad_trees are associated to tables by means of the ad_table_id column, instead of the treetype column.