Fixes issue 25827: Time fields work properly in timezones with negative offsets
Depending on who calls the formSaved function of OBTimeItem, the provided time can be already converted to the client local time or not. To know the time has a
lready been converted, the function used to check if the full year of the time was 1970. If that was the case, then the time was converted from UTC to local ti
me.

The problem is that if the user is working from a timezone with negative offset the full year will not be 1970, but 1969. This has been fixed by replacing data
[this.name].getFullYear() === 1970 with data[this.name].getFullYear() <= 1970