Fixes issue 25838: Time fields work properly in timezones with negative offsets
There was a problem the todays date was assigned to a datetime:

Original date: 31/12/2013 10:00:00
Todays date: 25/02/2014

The assignment order was year, month day, so the datetime was transformed like this:
Year updated: 31/12/2014 10:00:00
Month updated: 31/02/2014 does not exists, so it is converted to 03/03/2014
Day updated: 25/03/2014

This has been fixed by changing the updating order to Day, Month, Year. Before the update the date month is set to January to prevent setting a day not supported by the current month.

--HG--
extra : rebase_source : a46bc67c2565bc00cdb146b539dc51129f7cd73b