[xsql-scrolling] Implement reading of xsql results using a scrollable interface This implements a new optional feature to the Sqlc codegen for xsql files.
It can be activated per SqlMethod by using return="scrollable" instead of the
typtical return="multuple".

When used the interface for using this SqlMethod changes. Instead of it returning
an array of the generated class (i.e. ReportGeneralLedgerData[]) now it returns
a single instance of this class. However this class then implements a new interface
called 'ScrollableFieldProvider'.
This interface allows to read the result row by row without loading all of it into
memory avoiding huge memory usage and out of memory problems when processing a lot
of result rows.

Additionally this adds a new version of the renderJR-method which allows to pass
this data directly to jasper allowing streaming of data from xsql up to jasper
without loading all of the rows ever into memory.

Note: that Sqlc code-generation is completely unchanged for all code not explicitely
using the new feature.

As now not all results are available at the same tiem code which is looking at or
modifying such data before passing to to JasperReports needs to be adapted.

Included here is one helper class to implement a common case of checking number
of result rows and stopping processing is a limit is exceeded.
This is used in several records to comply with row-number restriction on xls export.

Common pattern to convert simple jasper call:
Before:
FieldProvider[] data = SomeDataClass.select (....);
renderJR(...., data, .... );

After:

ScrollableFieldProfier data = null;
try {
  data= SomeDataclass.select( .... );
  renderJR(..., data, ....);
} finally {
  if (data != null) {
    data.close();
}

IMPORTANT NOTE: It is very important to add this finally block which calls close() always as before calling close
a connection, transaction, result set for this query are kept open.

And all those must be close always (!)

The finally block allows to do this easily in case of success and error.

--HG--
extra : rebase_source : f4f332c8fc510dc6ecb5f339d30c7e5b226bb806