[Tree Extension] Use canBeParentNode instead of canAcceptDrop
Using canAcceptDrop to handle what nodes can be parent was not working properly, it was not possible to move a node to another position in its same parent if
its siblings had summaryLevel=false

This has been fixed by using a new property called canBeParentNode. The transferNodes function cancels the node movement is the new parent node can not be a p
arent node.