[Tree Extension] Implements addNodeCommonAttributes in TreeDatasourceService
There are some attributes that have to be added to all nodes, and that its value does not depend on the specific datasource implementation. To set this attributes the addNodeCommonAttributes method has been added to TreeDatasourceService. Whenever a subclass datasource implementation is going to build a json node, it must call to this method to have this attributes set to the node.

For now it is being used to set the canBeParentNode and the node icon.