Fixes issue 25610: text area fields do not parse properly the filter criteria
The problem was that the text area columns where not being considered text column when they were filtered. This was because they used the 'textArea' type, which was not included in the smartclientbuiltinTypes. This has been fixed by including it, and defining that the textAreaType inherits from the text type.

--HG--
extra : rebase_source : b3a61c8833ab6792ed64b6dfb7559c67c2194390