Fixes issue 25534: Problem when creating records in grouped grids
When a record was created in a grouped grid, the created record was not included in the grid until it was refreshed manually. Now it is updated automatically,
 its parent record is opened and the created record is scrolled to the middle of the grid.

Also, calling refreshCurrentRecord did not update the grouped grid, and closeProcessPopup did not update it either.

--HG--
extra : rebase_source : c6c6062f1afe8a415e1406b0755fc588f78c59b5