Fixes bug 25951:Auxiliary inputs are executed when FIC is called in CHANGE mode
Before this change, the FIC did not execute the auxiliary inputs when it was called in CHANGE mode. This was done to prevent resetting auxiliary inputs that ha
d been modified in callouts.

This made impossible to use auxiliary inputs in display logic, because they would not be updated until the records was refreshed. To fix this while allowing to
 overwrite auxiliary inputs from callouts, this has been done:
- The list of auxiliary inputs modified in callouts is persisted. This list is reset when the FIC is called in a mode other than CHANGE (originally in this cas
e all the auxiliary inputs were being updated, regardless of whether they had been modified by a callout).
- If the FIC is called in CHANGE mode, then only the auxiliary inputs that have not been modified by callouts are recalculated.
- The auxiliary inputs are now executed after the callouts are run, in order to be able to use the updated data in the auxiliary input calculation.

--HG--
extra : rebase_source : affe8ea1944ee31364c3d572950597aa66ac5732