Fixes issue 26008:  Problem while picking a business partner using a selector
The problem was that the overwrittenAuxiliaryInputs parameter (a string array) was being returned in the getContextInfo function, and sent to all action handlers. Some action handlers assume that the parameters do not contain arrays, and where having problems trying to convert the string array to a plain string.

The overwrittenAuxiliaryInputs parameter is only used in the FIC, and only when it is called in CHANGE mode. For all other modes, the paramater is reset to an empty string array in the FIC. To fix this, the overwrittenAuxiliaryInputs parameter is only sent to the FIC, and only in CHANGE mode.

--HG--
extra : rebase_source : fdd46d2e31f8b9df87acab2d0017cfc7bfa96b75