Fixes issue 25558: Fixes callout of manual forms
The code that applies the response of the manual forms callout was taking for granted that there one only one frame called appFrame. This was true in Openbravo 2.50, but in Openbravo 3.0 many manual windows can be opened at the same time, so there may be several frames called appFrame in one given moment.

This has been fixed by storing the active manual frame (the one whose tab is focused), and using that frame when applying the response of the callout.

--HG--
extra : rebase_source : b8732a1b0b4c638cb51ad965dbeb06611baeb830