Fixes issue 26011: Selector Out Field don't work in grid mode
The problem was that the fields[j].setValue function was being used to apply store the value of the out fields, and that function is not available when the sel
ector is used in the grid view. To solve this, grid.setEditValue is used if the fields[j].setValue function is not available.

--HG--
extra : rebase_source : 52289a0f23cab4ea78d81f592a73ad5812ffaae8