Fixes issue 26049: Filter on the product characteristics window is not working
The problem was that the Characteristic field of the Value subtab was being shown. That field act as link to the parent tab, and if it is shown the criteria used to fetch the subtab records could not be properly formed. This has been fixed by not showing the field in grid view.

--HG--
extra : transplant_source : %AD%1Bk%29%01F%C7%93%D4%25%ED%93%FE%3E%23%2Aq%B6r9